+++ hassio-addon/CHANGELOG.md
@@ -1,7 +1,7 @@
 v2.5.0
 - feat: 엔티티 이름 변경시 한글일 경우 entityId 를 로마자변환 추가 및 entityId 유지 옵션 추가. 
 - fix: 일부 엔티티의 discovery에서 없는 속성을 발행하는 문제 수정
-
+- fix: command_*가 script를 실행할때 args를 전달하지 않는 문제 수정
 
 v2.4.0
 - feat: packet_defaults에 rx_min_length, rx_max_length 옵션 도입